move Delegate to types.d.ts
