+++ src/lib/components/SpeakerList.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
     import DelLabel from "$lib/components/del-label/DelLabel.svelte";
     import DelPopup, { defaultPlaceholder, defaultPopupSettings } from "$lib/components/del-input/DelPopup.svelte";
-    import type { Delegate } from "$lib/db";
     import { findDelegate } from "$lib/db/del";
     import { formatValidationError } from "$lib/motions/form_validation";
-    import type { DelegateID, Speaker, SpeakerEntryID } from "$lib/types";
+    import type { Delegate, DelegateID, Speaker, SpeakerEntryID } from "$lib/types";
     import { getDndItemId, isDndShadow, processDrag } from "$lib/util/dnd";
     import { isPresent, triggerConfirmModal } from "$lib/util";
