+++ src/lib/motions/definitions.ts
@@ -1,9 +1,8 @@
-import type { Motion, MotionKind, SortEntry } from "$lib/types";
+import type { Delegate, Motion, MotionKind, SortEntry } from "$lib/types";
 import { nonEmptyString, presentDelegateSchema, refineSpeakingTime, timeSchema, topicSchema } from "$lib/motions/form_validation";
 import type { MotionInput } from "$lib/motions/types";
 import { z } from "zod";
 import { stringifyTime } from "$lib/util/time";
-import type { Delegate } from "$lib/db";
 import { findDelegate } from "$lib/db/del";
 
 /**
