+++ src/routes/admin/stats/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
     import MetaTags from "$lib/components/MetaTags.svelte";
     import DelLabel from "$lib/components/del-label/DelLabel.svelte";
-    import { db, type Delegate } from "$lib/db";
+    import { db } from "$lib/db";
     import { defaultStats, enabledDelegatesStore } from "$lib/db/del";
     import { getSessionDataContext } from "$lib/stores/session";
-    import type { StatsData } from "$lib/types";
+    import type { Delegate, StatsData } from "$lib/types";
     import { compare, downloadFile, isPresent, triggerConfirmModal } from "$lib/util";
     import { stringifyTime } from "$lib/util/time";
